+++ idlib/Parser.h
@@ -225,9 +225,11 @@ private:
 	int				Directive_ifndef( void );
 	int				Directive_else( void );
 	int				Directive_endif( void );
-	int				EvaluateTokens( idToken *tokens, signed long int *intvalue, double *floatvalue, int integer );
-	int				Evaluate( signed long int *intvalue, double *floatvalue, int integer );
-	int				DollarEvaluate( signed long int *intvalue, double *floatvalue, int integer);
+	// flibit: 64 bit fix, removed long keyword
+	int				EvaluateTokens( idToken *tokens, int *intvalue, double *floatvalue, int integer );
+	int				Evaluate( int *intvalue, double *floatvalue, int integer );
+	int				DollarEvaluate( int *intvalue, double *floatvalue, int integer);
+	// flibit end
 	int				Directive_define( void );
 	int				Directive_elif( void );
 	int				Directive_if( void );
