+++ idlib/hashing/CRC32.cpp
@@ -12,7 +12,9 @@
 
 #ifdef CREATE_CRC_TABLE
 
-static unsigned long crctable[256];
+// RB: 64 bit fix, changed long to int
+static unsigned int crctable[256];
+// RB end
 
 /*
    Generate a table for a byte-wise 32-bit CRC calculation on the polynomial:
@@ -41,7 +43,9 @@ static unsigned long crctable[256];
 
 void make_crc_table( void ) {
 	int i, j;
-	unsigned long c, poly;
+	// RB: 64 bit fix, changed long to int
+	unsigned int c, poly;
+	// RB end
 	/* terms of polynomial defining this crc (except x^32): */
 	static const byte p[] = {0,1,2,4,5,7,8,10,11,12,16,22,23,26};
 
@@ -52,7 +56,9 @@ void make_crc_table( void ) {
 	}
 
 	for ( i = 0; i < 256; i++ ) {
-		c = (unsigned long)i;
+		// RB: 64 bit fix, changed long to int
+		c = (unsigned int)i;
+		// RB end
 		for ( j = 0; j < 8; j++ ) {
 			c = ( c & 1 ) ? poly ^ ( c >> 1 ) : ( c >> 1 );
 		}
@@ -65,7 +71,9 @@ void make_crc_table( void ) {
 /*
   Table of CRC-32's of all single-byte values (made by make_crc_table)
 */
-static unsigned long crctable[256] = {
+// RB: 64 bit fix, changed long to int
+static unsigned int crctable[256] = {
+// RB end
 	0x00000000L, 0x77073096L, 0xee0e612cL, 0x990951baL,
 	0x076dc419L, 0x706af48fL, 0xe963a535L, 0x9e6495a3L,
 	0x0edb8832L, 0x79dcb8a4L, 0xe0d5e91eL, 0x97d2d988L,
@@ -134,16 +142,17 @@ static unsigned long crctable[256] = {
 
 #endif
 
-void CRC32_InitChecksum( unsigned long &crcvalue ) {
+// RB: 64 bit fixes, changed long to int
+void CRC32_InitChecksum( unsigned int &crcvalue ) {
 	crcvalue = CRC32_INIT_VALUE;
 }
 
-void CRC32_Update( unsigned long &crcvalue, const byte data ) {
+void CRC32_Update( unsigned int &crcvalue, const byte data ) {
 	crcvalue = crctable[ ( crcvalue ^ data ) & 0xff ] ^ ( crcvalue >> 8 );
 }
 
-void CRC32_UpdateChecksum( unsigned long &crcvalue, const void *data, int length ) {
-	unsigned long crc;
+void CRC32_UpdateChecksum( unsigned int &crcvalue, const void *data, int length ) {
+	unsigned int crc;
 	const unsigned char *buf = (const unsigned char *) data;
 
 	crc = crcvalue;
@@ -153,15 +162,16 @@ void CRC32_UpdateChecksum( unsigned long &crcvalue, const void *data, int length
 	crcvalue = crc;
 }
 
-void CRC32_FinishChecksum( unsigned long &crcvalue ) {
+void CRC32_FinishChecksum( unsigned int &crcvalue ) {
 	crcvalue ^= CRC32_XOR_VALUE;
 }
 
-unsigned long CRC32_BlockChecksum( const void *data, int length ) {
-	unsigned long crc;
+unsigned int CRC32_BlockChecksum( const void *data, int length ) {
+	unsigned int crc;
 
 	CRC32_InitChecksum( crc );
 	CRC32_UpdateChecksum( crc, data, length );
 	CRC32_FinishChecksum( crc );
 	return crc;
 }
+// RB end
